refactor(application): reorganize aspect management into dedicated packages
Move aspect-related use cases to dedicated subdirectories:
- command/aspect/: DeleteAspectDefinitionUseCase, UpdateAspectDefinitionUseCase
- query/aspect/: GetAspectDefinitionUseCase, ListAspectDefinitionsUseCase

Improves code organization and follows package-by-feature structure
consistent with Context View organization.
